requirements.txt
# --- Core numeric / IO
numpy==1.26.4
pandas==2.2.2
pyarrow>=16.1.0

# --- API
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings>=2.0.0
python-dotenv==1.0.1
requests==2.32.3
slowapi==0.1.9

# --- UI

watchdog==4.0.1

# --- Embeddings / NLP
sentence-transformers==3.0.1
transformers==4.43.3
tokenizers>=0.19.1
accelerate==0.33.0
tiktoken>=0.7.0

# PyTorch (install works on macOS CPU by default; use extra index https://download.pytorch.org/whl/cpu )
torch==2.8.0

# --- Vector search
faiss-cpu==1.12.0

# --- Dev / DX
rich==13.7.1
ruff==0.6.8
pytest==8.3.3
invoke==2.1.1
black==24.10.0

# --- YAML + LLM client
PyYAML==6.0.2
openai==1.47.0
langchain-openai==0.1.23
httpx==0.27.2


#----LangChain (0.3.x series for LangGraph compatibility)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langsmith>=0.2.0
langgraph>=0.2.0

# --- Azure & Database
azure-ai-documentintelligence>=1.0.0b3
psycopg2-binary==2.9.9
pgvector==0.2.5
sqlalchemy==2.0.30
